backend/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
numpy==1.22.1
PyJWT==1.7.1
SQLAlchemy==1.4.35